test(output): serilization
